requirements.txt
discord.py
python-twitch-client
aiohttp